-- Changed equipment enhancement to lose level on upgrade, with a crystal in donation shop to prevent the loss -- Slayer Armor added in with slayer helmets and upgrading -- Slayer reset added in -- fixed multiple claim problems with tracks -- fixed tarn tower npc on level 9 that wasnt supposed to be there -- Multiple packs adjusted (Donator Pack I , II, II, Ultimate Donator) -- Multiple packs added (Slayer casket, Vote Box, PvM Casket, PvM Locked Chest, Slayer Locked Chest, Vote Locked Chest (With Keys), Ultimate Cert Pack, Pet Enhancement Pack) -- Redid commands and fixed up old ones -- Global bosses set up with defs, drops, damage limits, and proper commands to reach them. (scripts will be redone soon, for now they share one) -- Fixed bug that showed pets in drop interface

+++ src/main/java/com/ruse/world/packages/combat/max/MagicMax.java
@@ -186,7 +186,7 @@ public class MagicMax {
                 maxHit *= 2;
             }
 
-            maxHit *= multiplyDamage(player);
+            maxHit *= (long) multiplyDamage(player);
 
             if (player.getTimers().get("double-damage") != null)
                 if (player.getTimers().get("double-damage").returnLeft() > 0) {
@@ -198,11 +198,11 @@ public class MagicMax {
 
             maxHit *= 10;
 
-            maxHit *= player.getEquipment().getBonus() == null ? 1 : player.getEquipment().getBonus().mageDamage();
+            maxHit *= player.getEquipment().getBonus() == null ? 1 : (long) player.getEquipment().getBonus().mageDamage();
         }
 
         if (victim != null && victim.isNpc() && (entity.isPlayer() && !entity.asPlayer().getRank().isDeveloper())) {
-            maxHit = (long) NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
+            maxHit = NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
         }
 
         return maxHit;
