+++ src/main/java/com/ruse/world/packages/combat/max/MeleeMax.java
@@ -347,7 +347,7 @@ public class MeleeMax {
                 maxHit *= 2;
             }
 
-            maxHit *= multiplyDamage(player);
+            maxHit *= (long) multiplyDamage(player);
 
             if (player.getTimers().get("double-damage") != null)
                 if (player.getTimers().get("double-damage").returnLeft() > 0) {
@@ -359,12 +359,12 @@ public class MeleeMax {
 
             maxHit *= 10;
 
-            maxHit *= player.getEquipment().getBonus() == null ? 1 : player.getEquipment().getBonus().meleeDamage();
+            maxHit *= player.getEquipment().getBonus() == null ? 1 : (long) player.getEquipment().getBonus().meleeDamage();
         }
 
 
         if (victim != null && victim.isNpc() && (entity.isPlayer() && !entity.asPlayer().getRank().isDeveloper())) {
-            maxHit = (long) NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
+            maxHit = NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
         }
 
         return maxHit;
