+++ src/main/java/com/ruse/world/packages/combat/max/RangeMax.java
@@ -183,7 +183,7 @@ public class RangeMax {
                 maxHit *= 2;
             }
 
-            maxHit *= multiplyDamage(player);
+            maxHit *= (long) multiplyDamage(player);
 
             if (player.getTimers().get("double-damage") != null)
                 if (player.getTimers().get("double-damage").returnLeft() > 0) {
@@ -195,12 +195,12 @@ public class RangeMax {
 
             maxHit *= 10;
 
-            maxHit *= player.getEquipment().getBonus() == null ? 1 : player.getEquipment().getBonus().rangeDamage();
+            maxHit *= player.getEquipment().getBonus() == null ? 1 : (long) player.getEquipment().getBonus().rangeDamage();
         }
 
 
         if (victim != null && victim.isNpc() && (entity.isPlayer() && !entity.asPlayer().getRank().isDeveloper())) {
-            maxHit = (long) NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
+            maxHit = NpcMaxHitLimit.limit((NPC) victim, maxHit, entity.asPlayer());
         }
 
         return maxHit;
